+++ Source/BE/src/main/java/imitori/mongodb/repository/JAENDicMonRepository.java
@@ -14,13 +14,13 @@ public class JAENDicMonRepository {
     @Autowired
     MongoTemplate mongoTemplate;
 
-    public long getMaxWordId() {
+    public Integer getMaxWordId() {
         Query query = new Query();
         query.with(new Sort(Sort.Direction.DESC, "id"));
         query.limit(1);
         JAENDicMonEntity maxObject = mongoTemplate.findOne(query, JAENDicMonEntity.class);
         if (maxObject == null) {
-            return 0L;
+            return 0;
         }
         return maxObject.getid();
     }
@@ -29,7 +29,7 @@ public class JAENDicMonRepository {
         return mongoTemplate.findAll(JAENDicMonEntity.class).size();
     }
 
-    public void insertWord(long w) {
+    public void insertWord(Integer w) {
         JAENDicMonEntity word = new JAENDicMonEntity();
         word.setent_seq(w);
